compiler: extract from ast_parse
