+++ CMakeLists.txt
@@ -785,6 +785,7 @@ include/daScript/misc/job_que.h
 include/daScript/misc/handle_registry.h
 include/daScript/misc/uric.h
 include/daScript/misc/lexer_alloc_track.h
+src/misc/das_common.cpp
 src/misc/gc_node.cpp
 src/misc/globals.cpp
 src/misc/sysos.cpp
@@ -1286,7 +1287,6 @@ SET(DAS_RELEASE_INCLUDE
     include/daScript/daScript.h
     include/daScript/daScriptBind.h
     include/daScript/daScriptModule.h
-    include/daScript/das_common.h
     include/daScript/das_config.h
     include/daScript/das_project_specific.h
 )
@@ -1296,6 +1296,7 @@ SET(DAS_RELEASE_MISC_INCLUDE
     include/daScript/misc/arraytype.h
     include/daScript/misc/callable.h
     include/daScript/misc/crash_handler.h
+    include/daScript/misc/das_common.h
     include/daScript/misc/debug_break.h
     include/daScript/misc/enums.h
     include/daScript/misc/fpe.h
