+++ utils/daScript/main.cpp
@@ -3,7 +3,7 @@
 #include "daScript/ast/dyn_modules.h"
 #include "daScript/daScript.h"
 #include "daScript/daScriptModule.h"
-#include "daScript/das_common.h"
+#include "daScript/misc/das_common.h"
 #include "daScript/simulate/fs_file_info.h"
 #include "../dasFormatter/fmt.h"
 #include "daScript/ast/ast_aot_cpp.h"
